added render functions

+++ app/controllers/api/v1/videos_controller.rb
@@ -1,13 +1,13 @@
 module Api
   module V1
     class VideosController < ApplicationController
-      before_action :set_video, only: %i[ show update destroy ]
-      before_action :get_meeting, only: %i[ index create show update destroy ]
+      before_action :set_video, only: %i[ show destroy render_status ]
+      before_action :get_meeting, only: %i[ index create show update destroy render_meeting ]
       before_action :get_user, only: %i[ index create show update destroy ]
 
       # GET /videos
       def index
-        @videos = @meeting.videos
+        @videos = @meeting.videos.sort
 
         render json: @videos
       end
@@ -51,7 +51,7 @@ module Api
           @video.update(
             video_type: 'user upload',
             upload_status: status[:data]['ingest']['status'],
-            duration: status[:data]['encoding']['metadata']['duration'],
+            duration: status[:data]['encoding']['metadata']['duration'].to_i,
             embed_url: upload[:data]['assets']['player'],
             source_url: upload[:data]['assets']['mp4'],
             api_video_id: upload[:data]['videoId']
@@ -65,16 +65,46 @@ module Api
       # DELETE /videos/1
       def destroy
         @video.destroy
-        auth = ApiVideo::Client.auth
-        delete = ApiVideo::Client.delete(auth[:data]['access_token'], @video['api_video_id'])
-      
+
+        if @video['render_id'] == nil
+          auth = ApiVideo::Client.auth
+          delete = ApiVideo::Client.delete(auth[:data]['access_token'], @video['api_video_id'])
+        end
+
         render json: {
           message: "Successfully deleted video."
         }, status: :ok
       end
 
       def render_meeting
+        @videos = @meeting.videos.map {|video| video}
+        @video = @meeting.videos.new(video_params)
+        
+        render_meeting = Shotstack::Client.render_meeting(@videos.sort)
 
+        if @video.save
+          @video.update(
+            video_type: 'rendered meeting',
+            render_status: 'rendering',
+            render_id: render_meeting[:data]['response']['id']
+          )
+          render json: @video, status: :created
+        else
+          render json: @video.errors, status: :unprocessable_entity
+        end
+      end
+
+      def render_status
+        if @video['upload_status'] != 'done'
+          render_status = Shotstack::Client.render_status(@video['render_id'])
+          @video.update(
+            render_status: render_status[:data]['response']['status'],
+            duration: render_status[:data]['response']['duration'].to_i,
+            source_url: render_status[:data]['response']['url']
+          )
+        end
+
+        render json: @video
       end
 
       private
